feat: Add imports, modules
Add syntax and functionality for importing code from other files.
The syntax is `import { ids... } from "./path/to/file";
For now, no transitive importing is done. This unfortunately causes
a not being able to interact with any type that a function returns,
unless that type is explicitly imported also.

This is sort of POC, and will probably be reworked further.